requirements.txt
blinker==1.9.0
click==8.3.1
epam-indigo==1.36.1
Flask==3.1.2
flask-cors==6.0.2
flask-swagger-ui==5.21.0
gunicorn==24.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
Werkzeug==3.1.5
